+++ code/engine.vc2008/xrGame/Explosive.cpp
@@ -182,7 +182,7 @@ struct SExpQParams
 ICF static BOOL grenade_hit_callback(collide::rq_result& result, LPVOID params)
 {
 	SExpQParams& ep	= *(SExpQParams*)params;
-	u16 mtl_idx			= GAMEMTL_NONE_IDX;
+	size_t mtl_idx	= GAMEMTL_NONE_IDX;
 	if(result.O){
 		IKinematics* V  = 0;
 		if (0!=(V=smart_cast<IKinematics*>(result.O->Visual()))){
